Give one module the words for a bearing
`compassWords` has lived inside `MeasuredToday` since the air card was
written, private to the component that prints "8 mph from the north-west".
The compass on the shore map needs the same words for its two needles, and
a component exporting a helper would make the map depend on the card's file.

Moved unchanged, so the card's behaviour is identical and its own test --
every compass point with its own words, plus the 359 wrap -- still covers
it end to end. The new tests cover the function directly and add the bucket
boundary the component test does not reach: 22 is north and 23 is
north-east, which is where an off-by-one would show and where rounding
through the middle of a bucket cannot.

The module records why eight points and not the sixteen the design brief's
example name uses. The two vocabularies would be on one screen -- "west" on
the card and "west-northwest" on a needle a few hundred pixels away, about
the same wind -- and the precision the finer rose was for is carried by the
degrees, which every caller states beside the words.
